Desafio concluido

+++ MinimalApi/Controllers/OrderController.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Mvc;
 using MinimalApi.Domains;
 using MinimalApi.Services;
+using MinimalApi.ViewModels;
 using MongoDB.Driver;
 
 namespace MinimalApi.Controllers
@@ -12,10 +13,14 @@ namespace MinimalApi.Controllers
     public class OrderController : ControllerBase
     {
         private readonly IMongoCollection<Order> _order;
+        private readonly IMongoCollection<Client> _client;
+        private readonly IMongoCollection<Product> _product;
 
         public OrderController(MongoDbService mongoDbService)
         {
             _order = mongoDbService.GetDatabase.GetCollection<Order>("Order");
+            _client = mongoDbService.GetDatabase.GetCollection<Client>("Client");
+            _product = mongoDbService.GetDatabase.GetCollection<Product>("product");
         }
 
         [HttpGet]
@@ -23,7 +28,8 @@ namespace MinimalApi.Controllers
         {
             try
             {
-                return Ok(await _order.Find(FilterDefinition<Order>.Empty).ToListAsync());
+                var res = _order.Find(FilterDefinition<Order>.Empty).ToList();
+                return Ok(res);
             }
             catch (Exception e)
             {
@@ -31,15 +37,45 @@ namespace MinimalApi.Controllers
             }
         }
 
-
         [HttpPost]
-        public async Task<ActionResult<Order>> Post(Order order)
+        public async Task<ActionResult<Order>> Post(OrderViewModel order)
         {
             try
             {
-                await _order.InsertOneAsync(order);
+                Order order1 = new Order();
+                order1.Id = order.Id;
+                order1.ProductId = order.ProductId;
+                order1.ClientId = order.ClientId;
+                order1.Status = order.Status;
+                order1.Date = order.Date;
 
-                return Ok();
+                Client client = await _client.Find(x => x.Id == order.ClientId).FirstOrDefaultAsync();
+
+                if (client == null)
+                {
+                    return NotFound("O Id do Cliente não é válido");
+                }
+
+                order1.Client = client;
+
+
+
+                order1.Products = new List<Product>();
+
+                foreach (string id in order1.ProductId!)
+                {
+                    Product product = await _product.Find(x => x.Id == id).FirstOrDefaultAsync();
+
+                    if (product != null)
+                    {
+                        order1.Products.Add(product);
+                    }
+
+                }
+
+                await _order.InsertOneAsync(order1);
+
+                return StatusCode(201);
             }
             catch (Exception e)
             {
@@ -52,10 +88,14 @@ namespace MinimalApi.Controllers
         {
             try
             {
-                var filter = Builders<Order>.Filter.Eq(x => x.Id, id);
-                var obj = await _order.FindAsync(filter);
+                var obj = await _order.Find(x => x.Id == id).FirstOrDefaultAsync();
 
-                return Ok(obj.First());
+                if (obj == null)
+                {
+                    return NotFound("O Id do pedido não foi encontrado !!!");
+                }
+
+                return Ok(obj);
             }
             catch (Exception e)
             {
@@ -64,13 +104,58 @@ namespace MinimalApi.Controllers
         }
 
         [HttpPut]
-        public async Task<ActionResult> Put(Order order)
+        public async Task<ActionResult> Put(Order order, string orderId)
         {
             try
             {
-                await _order.ReplaceOneAsync(x => x.Id == order.Id, order);
 
-                return Ok("Atualizado com sucesso");
+                var orderExists = await _order.Find(x => x.Id == orderId).FirstOrDefaultAsync();
+
+                if (orderExists != null)
+                {
+
+                    Order order1 = await _order.Find(x => x.Id == orderId).FirstOrDefaultAsync();
+                    order1.ProductId = order.ProductId != null ? order.ProductId : order1.ProductId;
+                    order1.ClientId = order.ClientId != null ? order.ClientId : order1.ClientId;
+                    order1.Status = order.Status != null ? order.Status : order1.Status;
+                    order1.Date = order.Date != null ? order.Date : order1.Date;
+
+                    if (order.ClientId != null)
+                    {
+                        Client client = await _client.Find(x => x.Id == order.ClientId).FirstOrDefaultAsync();
+
+                        if (client == null)
+                        {
+                            return NotFound("O Id do cliente não é válido");
+                        }
+
+                        order1.Client = client;
+                    }
+
+                    if (order.ProductId != null)
+                    {
+                        var updateProdutos = new List<Product>();
+
+                        foreach (string id in order.ProductId)
+                        {
+                            Product product = await _product.Find(x => x.Id == id).FirstOrDefaultAsync();
+
+                            if (product != null)
+                            {
+                                updateProdutos.Add(product);
+                            }
+                        }
+
+                        order1.Products = updateProdutos;
+                    }
+
+
+                    await _order.ReplaceOneAsync(x => x.Id == orderId, order1);
+
+                    return Ok("Atualizado com sucesso");
+                }
+
+                return NotFound("Não foi encontrado nenhum pedido com esse Id !!!");
             }
             catch (Exception e)
             {
@@ -83,6 +168,12 @@ namespace MinimalApi.Controllers
         {
             try
             {
+
+                if ((await _order.Find(x => x.Id == id).FirstOrDefaultAsync()) == null)
+                {
+                    return NotFound("O Id do pedido não foi encontrado !!!");
+                }
+
                 await _order.DeleteOneAsync(x => x.Id == id);
 
                 return NoContent();
